# app/src/main/resources/openapi.json
{
  "openapi": "3.0.1",
  "info": {
    "title": "User Management System",
    "description": "Create and Manage Users, Roles and Permissions",
    "contact": {
      "name": "Bibek Aryal"
    },
    "license": {
      "name": "Personal Use Only"
    },
    "version": "1.0.1"
  },
  "servers": [
    {
      "url": "http://localhost:8080/ums",
      "description": "Generated server url"
    }
  ],
...
  "paths": {
...
    "/api/v1/app_roles/role/{id}": {
      "get": {
        "tags": [
          "Roles Management"
        ],
        "summary": "Get role by ID",
        "description": "Retrieves a specific role by its unique ID",
        "operationId": "readAppRole",
        "parameters": [
          {
            "name": "id",
            "in": "path",
            "description": "ID of the role to retrieve",
            "required": true,
            "schema": {
              "type": "integer",
              "format": "int32"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Role retrieved successfully",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/AppRoleResponse"
                }
              }
            }
          },
          "401": {
            "description": "Unauthorized - User Not Authorized",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/AppRoleResponse"
                }
              }
            }
          },
          "403": {
            "description": "Forbidden - User Forbidden or Not Validated",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/AppRoleResponse"
                }
              }
            }
          },
          "404": {
            "description": "Not Found - Role Not Found",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/AppRoleResponse"
                }
              }
            }
          },
          "500": {
            "description": "Internal Server Error - Other Errors",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/AppRoleResponse"
                }
              }
            }
          }
        },
        "security": [
          {
            "Token": []
          }
        ]
      },
      "put": {
        "tags": [
          "Roles Management"
        ],
        "summary": "Update a role by ID",
        "description": "Updates an existing role identified by its unique ID",
        "operationId": "updateAppRole",
        "parameters": [
          {
            "name": "id",
            "in": "path",
            "description": "ID of the role to update",
            "required": true,
            "schema": {
              "type": "integer",
              "format": "int32"
            }
          }
        ],
        "requestBody": {
          "description": "Details of the role to update",
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/AppRoleRequest"
              }
            }
          },
          "required": true
        },
        "responses": {
          "200": {
            "description": "Role updated successfully",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/AppRoleResponse"
                }
              }
            }
          },
          "400": {
            "description": "Bad Request - Required Element Missing",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ResponseStatusInfo"
                }
              }
            }
          },
          "401": {
            "description": "Unauthorized - User Not Authorized",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/AppRoleResponse"
                }
              }
            }
          },
          "403": {
            "description": "Forbidden - User Forbidden or Not Validated",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/AppRoleResponse"
                }
              }
            }
          },
          "404": {
            "description": "Not Found - Role Not Found",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/AppRoleResponse"
                }
              }
            }
          },
          "500": {
            "description": "Internal Server Error - Other Errors",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/AppRoleResponse"
                }
              }
            }
          }
        },
        "security": [
          {
            "Token": []
          }
        ]
      },
      "delete": {
        "tags": [
          "Roles Management"
        ],
        "summary": "Soft delete a role by ID",
        "description": "Marks a role as deleted without permanently removing it, identified by its unique ID",
        "operationId": "softDeleteAppRole",
        "parameters": [
          {
            "name": "id",
            "in": "path",
            "description": "ID of the role to soft delete",
            "required": true,
            "schema": {
              "type": "integer",
              "format": "int32"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Role soft deleted successfully",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/AppRoleResponse"
                }
              }
            }
          },
          "401": {
            "description": "Unauthorized - User Not Authorized",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/AppRoleResponse"
                }
              }
            }
          },
          "403": {
            "description": "Forbidden - User Forbidden or Not Validated",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/AppRoleResponse"
                }
              }
            }
          },
          "404": {
            "description": "Not Found - Role Not Found",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/AppRoleResponse"
                }
              }
            }
          },
          "500": {
            "description": "Internal Server Error - Other Errors",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/AppRoleResponse"
                }
              }
            }
          }
        },
        "security": [
          {
            "Token": []
          }
        ]
      }
    },
...
    "/api/v1/app_roles_permissions/role_permission": {
      "post": {
        "tags": [
          "Roles Permissions Management"
        ],
        "summary": "Assign Permission to a Role",
        "description": "Adds a Permission to a Role",
        "operationId": "createAppRolePermission",
        "requestBody": {
          "description": "Details of permission and role to be linked",
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/AppRolePermissionRequest"
              }
            }
          },
          "required": true
        },
        "responses": {
          "200": {
            "description": "Permission assigned successfully to role",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/AppRolePermissionResponse"
                }
              }
            }
          },
          "400": {
            "description": "Bad Request - Required Element Missing",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ResponseStatusInfo"
                }
              }
            }
          },
          "401": {
            "description": "Unauthorized - User Not Authorized",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/AppRolePermissionResponse"
                }
              }
            }
          },
          "403": {
            "description": "Forbidden - User Forbidden or Not Validated",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/AppRolePermissionResponse"
                }
              }
            }
          },
          "404": {
            "description": "Not Found - Role or Permission to Link Not Found",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/AppRolePermissionResponse"
                }
              }
            }
          },
          "500": {
            "description": "Internal Server Error - Other Errors",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/AppRolePermissionResponse"
                }
              }
            }
          }
        },
        "security": [
          {
            "Token": []
          }
        ]
      }
    },
...
    "/api/v1/app_roles_permissions": {
      "get": {
        "tags": [
          "Roles Permissions Management"
        ],
        "summary": "Get all roles and permissions linked",
        "description": "Retrieves all linked roles and permissions available in the system",
        "operationId": "readAppRolePermissions",
        "responses": {
          "200": {
            "description": "Roles Permissions retrieved successfully",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/AppRolePermissionResponse"
                }
              }
            }
          },
          "401": {
            "description": "Unauthorized - User Not Authorized",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/AppRolePermissionResponse"
                }
              }
            }
          },
          "403": {
            "description": "Forbidden - User Forbidden or Not Validated",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/AppRolePermissionResponse"
                }
              }
            }
          },
          "500": {
            "description": "Internal Server Error - Other Errors",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/AppRolePermissionResponse"
                }
              }
            }
          }
        },
        "security": [
          {
            "Token": []
          }
        ]
      }
    },
    "/api/v1/app_roles_permissions/role_permission/{roleId}/{permissionId}": {
      "get": {
        "tags": [
          "Roles Permissions Management"
        ],
        "summary": "Get a linked role permission role by role ID and permission ID",
        "description": "Retrieves a specific role permission by its role ID and permission ID",
        "operationId": "readAppRolePermission",
        "parameters": [
          {
            "name": "roleId",
            "in": "path",
            "description": "ID of the role in linked role permission",
            "required": true,
            "schema": {
              "type": "integer",
              "format": "int32"
            }
          },
          {
            "name": "permissionId",
            "in": "path",
            "description": "ID of the permission in linked role permission",
            "required": true,
            "schema": {
              "type": "integer",
              "format": "int32"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Role Permission retrieved successfully",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/AppRolePermissionResponse"
                }
              }
            }
          },
          "401": {
            "description": "Unauthorized - User Not Authorized",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/AppRolePermissionResponse"
                }
              }
            }
          },
          "403": {
            "description": "Forbidden - User Forbidden or Not Validated",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/AppRolePermissionResponse"
                }
              }
            }
          },
          "404": {
            "description": "Not Found - Role Permission Not Found",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/AppRolePermissionResponse"
                }
              }
            }
          },
          "500": {
            "description": "Internal Server Error - Other Errors",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/AppRolePermissionResponse"
                }
              }
            }
          }
        },
        "security": [
          {
            "Token": []
          }
        ]
      },
      "delete": {
        "tags": [
          "Roles Permissions Management"
        ],
        "summary": "Unassign Permission from a Role",
        "description": "Removes a Permission from a Role",
        "operationId": "deleteAppRolePermission",
        "parameters": [
          {
            "name": "roleId",
            "in": "path",
            "description": "ID of the role in linked role permission",
            "required": true,
            "schema": {
              "type": "integer",
              "format": "int32"
            }
          },
          {
            "name": "permissionId",
            "in": "path",
            "description": "ID of the permission in linked role permission",
            "required": true,
            "schema": {
              "type": "integer",
              "format": "int32"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Permission unassigned successfully from role",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/AppRolePermissionResponse"
                }
              }
            }
          },
          "401": {
            "description": "Unauthorized - User Not Authorized",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/AppRolePermissionResponse"
                }
              }
            }
          },
          "403": {
            "description": "Forbidden - User Forbidden or Not Validated",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/AppRolePermissionResponse"
                }
              }
            }
          },
          "404": {
            "description": "Not Found - Role Permission Not Found",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/AppRolePermissionResponse"
                }
              }
            }
          },
          "500": {
            "description": "Internal Server Error - Other Errors",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/AppRolePermissionResponse"
                }
              }
            }
          }
        },
        "security": [
          {
            "Token": []
          }
        ]
      }
    },
    "/api/v1/app_roles_permissions/role/{roleId}": {
      "get": {
        "tags": [
          "Roles Permissions Management"
        ],
        "summary": "Get linked roles and permissions by role ID",
        "description": "Retrieves all roles and permissions associated with the specified role ID",
        "operationId": "readAppRolePermissionsByRoleId",
        "parameters": [
          {
            "name": "roleId",
            "in": "path",
            "description": "ID of the role for which roles permissions are retrieved",
            "required": true,
            "schema": {
              "type": "integer",
              "format": "int32"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Roles Permissions retrieved successfully",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/AppRolePermissionResponse"
                }
              }
            }
          },
          "401": {
            "description": "Unauthorized - User Not Authorized",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/AppRolePermissionResponse"
                }
              }
            }
          },
          "403": {
            "description": "Forbidden - User Forbidden or Not Validated",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/AppRolePermissionResponse"
                }
              }
            }
          },
          "500": {
            "description": "Internal Server Error - Other Errors",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/AppRolePermissionResponse"
                }
              }
            }
          }
        },
        "security": [
          {
            "Token": []
          }
        ]
      }
    },
    "/api/v1/app_roles_permissions/app/{appId}/roles/{roleIds}": {
      "get": {
        "tags": [
          "Roles Permissions Management"
        ],
        "summary": "Get linked roles and permissions by role ID and app ID",
        "description": "Retrieves all roles and permissions associated with the specified role ID for a given application ID",
        "operationId": "readAppRolePermissionsByRoleIds",
        "parameters": [
          {
            "name": "appId",
            "in": "path",
            "description": "ID of the application for which roles permissions are retrieved",
            "required": true,
            "schema": {
              "type": "string"
            }
          },
          {
            "name": "roleIds",
            "in": "path",
            "description": "IDs of the roles for which roles permissions are retrieved",
            "required": true,
            "schema": {
              "type": "array",
              "items": {
                "type": "integer",
                "format": "int32"
              }
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Roles Permissions retrieved successfully",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/AppRolePermissionResponse"
                }
              }
            }
          },
          "401": {
            "description": "Unauthorized - User Not Authorized",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/AppRolePermissionResponse"
                }
              }
            }
          },
          "403": {
            "description": "Forbidden - User Forbidden or Not Validated",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/AppRolePermissionResponse"
                }
              }
            }
          },
          "500": {
            "description": "Internal Server Error - Other Errors",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/AppRolePermissionResponse"
                }
              }
            }
          }
        },
        "security": [
          {
            "Token": []
          }
        ]
      }
    },
...
  },
  "components": {
    "schemas": {
...
      "ResponseCrudInfo": {
        "type": "object",
        "properties": {
          "insertedRowsCount": {
            "type": "integer",
            "format": "int32"
          },
          "updatedRowsCount": {
            "type": "integer",
            "format": "int32"
          },
          "deletedRowsCount": {
            "type": "integer",
            "format": "int32"
          },
          "restoredRowsCount": {
            "type": "integer",
            "format": "int32"
          }
        }
      },
      "ResponsePageInfo": {
        "type": "object",
        "properties": {
          "totalItems": {
            "type": "integer",
            "format": "int32"
          },
          "totalPages": {
            "type": "integer",
            "format": "int32"
          },
          "pageNumber": {
            "type": "integer",
            "format": "int32"
          },
          "perPage": {
            "type": "integer",
            "format": "int32"
          }
        }
      },
      "ResponseStatusInfo": {
        "type": "object",
        "properties": {
          "message": {
            "type": "string"
          },
          "errMsg": {
            "type": "string"
          }
        }
      },
...
      "AppPermissionDto": {
        "required": [
          "description",
          "name"
        ],
        "type": "object",
        "properties": {
          "name": {
            "type": "string"
          },
          "description": {
            "type": "string"
          },
          "id": {
            "type": "integer",
            "format": "int32"
          },
          "appId": {
            "type": "string"
          }
        }
      },
      "AppRoleDto": {
        "required": [
          "description",
          "name"
        ],
        "type": "object",
        "properties": {
          "name": {
            "type": "string"
          },
          "description": {
            "type": "string"
          },
          "id": {
            "type": "integer",
            "format": "int32"
          },
          "permissions": {
            "type": "array",
            "items": {
              "$ref": "#/components/schemas/AppPermissionDto"
            }
          }
        }
      },
...
      "AppRoleRequest": {
        "required": [
          "description",
          "name"
        ],
        "type": "object",
        "properties": {
          "name": {
            "type": "string"
          },
          "description": {
            "type": "string"
          }
        }
      },
      "AppRoleResponse": {
        "type": "object",
        "properties": {
          "responseCrudInfo": {
            "$ref": "#/components/schemas/ResponseCrudInfo"
          },
          "responsePageInfo": {
            "$ref": "#/components/schemas/ResponsePageInfo"
          },
          "responseStatusInfo": {
            "$ref": "#/components/schemas/ResponseStatusInfo"
          },
          "roles": {
            "type": "array",
            "items": {
              "$ref": "#/components/schemas/AppRoleDto"
            }
          }
        }
      },
...
      "AppRolePermissionRequest": {
        "type": "object",
        "properties": {
          "roleId": {
            "type": "integer",
            "format": "int32"
          },
          "permissionId": {
            "type": "integer",
            "format": "int32"
          }
        }
      },
      "AppRolePermissionDto": {
        "type": "object",
        "properties": {
          "role": {
            "$ref": "#/components/schemas/AppRoleDto"
          },
          "permission": {
            "$ref": "#/components/schemas/AppPermissionDto"
          }
        }
      },
      "AppRolePermissionResponse": {
        "type": "object",
        "properties": {
          "responseCrudInfo": {
            "$ref": "#/components/schemas/ResponseCrudInfo"
          },
          "responsePageInfo": {
            "$ref": "#/components/schemas/ResponsePageInfo"
          },
          "responseStatusInfo": {
            "$ref": "#/components/schemas/ResponseStatusInfo"
          },
          "rolesPermissions": {
            "type": "array",
            "items": {
              "$ref": "#/components/schemas/AppRolePermissionDto"
            }
          }
        }
      },
...
    },
    "securitySchemes": {
      "Token": {
        "type": "http",
        "scheme": "Bearer",
        "bearerFormat": "JWT"
      },
      "Basic": {
        "type": "http",
        "scheme": "Basic"
      }
    }
  }
}